backend/exercise1-api/requirements.txt
# Python依赖文件 - Exercise 1 API测试和示例程序

# HTTP请求库
requests>=2.28.0

# PostgreSQL数据库连接 (用于数据库测试)
psycopg2-binary>=2.9.0

# 可选: 更好的命令行输出
colorama>=0.4.0

# 可选: 进度条显示
tqdm>=4.64.0